moved make_copy to core.tree with respective tests. example deleted
